requirements.txt
fastapi==0.115.12
matplotlib==3.10.3
numpy==1.25.2
opencv_python==4.11.0.86
Pillow==11.2.1
pydantic==2.11.5
setuptools==75.8.0
SQLAlchemy==2.0.38
torch==2.6.0
torchmetrics==1.7.1
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
websockets==15.0.1
